api/requirements.txt
litestar>=2.8.3
asyncpg>=0.29.0
langchain>=0.2.16
langchain-community>=0.2.16
langchain-postgres>=0.0.12
sentence-transformers>=3.0.1
PyPDF2>=3.0.1
pdfplumber>=0.11.4
pandas>=2.2.3
beautifulsoup4>=4.12.3
httpx>=0.27.0
python-multipart>=0.0.9
pgvector>=0.2.5
python-dotenv>=1.0.1
uvicorn>=0.30.6
langchain-text-splitters>=1.1.0
langchain-ollama>=0.2.0
playwright>=1.57.0
pytesseract>=0.3.10
pdf2image>=1.17.0
rank-bm25>=0.2.2
diskcache>=5.6.3
langchain-huggingface>=0.1.0